requirements.txt
numpy>=1.21.0
PyQt5>=5.15.0
PyOpenGL>=3.1.9
PyOpenGL_accelerate>=3.1.9
pygame>=2.6.1
PyGLM>=2.8.1
matplotlib>=3.4.0
scipy>=1.10.0 